requirements.txt
alembic==1.5.4
aniso8601==8.1.1
apache-airflow==2.0.1
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.0
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-sqlite==1.0.1
apispec==3.3.2
argcomplete==1.12.2
attrs==20.3.0
Babel==2.9.0
bleach==3.2.1
cached-property==1.5.2
cachelib==0.1.1
cattrs==1.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.7.2
commonmark==0.9.1
connexion==2.7.0
croniter==0.3.37
cryptography==3.4.3
Cython==0.29.21
defusedxml==0.6.0
dill==0.3.3
dnspython==2.1.0
docutils==0.16
email-validator==1.1.2
Flask==1.1.2
Flask-AppBuilder==3.1.1
Flask-Babel==1.0.0
Flask-Caching==1.9.0
Flask-JWT-Extended==3.25.0
Flask-JWT-Simple==0.0.3
Flask-Login==0.4.1
Flask-OpenID==1.2.5
flask-restx==0.2.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.36.0
Flask-WTF==0.14.3
graphviz==0.16
greenlet==1.0.0
gunicorn==19.10.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.14
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
keyring==21.5.0
kosher==0.1.6
lazy-object-proxy==1.5.2
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.1.0
marshmallow-sqlalchemy==0.23.1
natsort==7.1.1
numpy==1.20.0
openapi-spec-validator==0.2.9
packaging==20.8
pandas==1.2.1
pendulum==2.1.2
pkginfo==1.6.1
prison==0.1.3
protobuf==3.14.0
protobuf-serialization==0.1.2
psutil==5.8.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
Pygments==2.7.3
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
readme-renderer==28.0
redis==3.5.3
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
rich==9.2.0
setproctitle==1.2.2
six==1.15.0
SQLAlchemy==1.4.0b2
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.36.8
swagger-ui-bundle==0.0.8
tabulate==0.8.7
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
twine==3.2.0
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.3
waitress==1.4.4
webencodings==0.5.1
Werkzeug==0.16.1
WTForms==2.3.3
zipp==3.4.0
